codigo com erro

+++ Back/src/main/java/com/example/Back/Entity/Matricula.java
@@ -14,7 +14,7 @@ public class Matricula {
 
     @Id
     @Column(name = "aluno_id")
-    private Integer alunoId;
+    private Long alunoId;
 
     @Id
     @Column(name = "curso_id")
